fix: 適切な場所に移動・statのコードを整理しました
