+++ lib/ui/screen/post_detail/component/post_detail_list_item.dart
@@ -16,10 +16,10 @@ import 'package:food_gram_app/router/go_router_extension.dart';
 import 'package:food_gram_app/router/router.dart';
 import 'package:food_gram_app/ui/component/app_convertible_price_text.dart';
 import 'package:food_gram_app/ui/component/app_elevated_button.dart';
+import 'package:food_gram_app/ui/component/app_profile_image.dart';
 import 'package:food_gram_app/ui/component/app_translatable_text.dart';
 import 'package:food_gram_app/ui/component/dialog/app_share_dialog.dart';
 import 'package:food_gram_app/ui/component/modal_sheet/save_album_picker_sheet.dart';
-import 'package:food_gram_app/ui/component/profile/app_profile_image.dart';
 import 'package:food_gram_app/ui/screen/post_detail/post_detail_view_model.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
