+++ lib/ui/screen/restaurant_review/restaurant_review_screen.dart
@@ -8,8 +8,8 @@ import 'package:food_gram_app/core/supabase/post/repository/map_post_repository.
 import 'package:food_gram_app/core/theme/style/restaurant_review_style.dart';
 import 'package:food_gram_app/gen/strings.g.dart';
 import 'package:food_gram_app/router/router.dart';
+import 'package:food_gram_app/ui/component/app_profile_image.dart';
 import 'package:food_gram_app/ui/component/common/app_async_value_group.dart';
-import 'package:food_gram_app/ui/component/profile/app_profile_image.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
